Fix new code to adhere to refactor
