+++ webapp/src/components/ParcelDetailPage/ParcelDetailPage.container.js
@@ -1,13 +1,16 @@
 import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
 
-import { getParams } from 'modules/location/reducer'
-import { getWallet } from 'modules/wallet/reducer'
-import { isLoading as isAddressLoading } from 'modules/address/reducer'
-import { getParcels, getError as getParcelError } from 'modules/parcels/reducer'
-import { getDistricts } from 'modules/districts/reducer'
-import { fetchWallet } from 'modules/wallet/actions'
-import { fetchParcel } from 'modules/parcels/actions'
+import { getParams } from 'modules/location/selectors'
+import { getWallet } from 'modules/wallet/selectors'
+import {
+  getParcels,
+  getError as getParcelError
+} from 'modules/parcels/selectors'
+import { isLoading as isAddressLoading } from 'modules/address/selectors'
+import { getDistricts } from 'modules/districts/selectors'
+import { fetchWalletRequest } from 'modules/wallet/actions'
+import { fetchParcelRequest } from 'modules/parcels/actions'
 import { openModal } from 'modules/ui/actions'
 import { navigateTo } from 'modules/location/actions'
 
@@ -33,8 +36,8 @@ const mapState = (state, ownProps) => {
 
 const mapDispatch = dispatch => ({
   onNavigate: location => dispatch(navigateTo(location)),
-  onConnect: () => dispatch(fetchWallet()),
-  onFetchParcel: (x, y) => dispatch(fetchParcel(x, y)),
+  onConnect: () => dispatch(fetchWalletRequest()),
+  onFetchParcel: (x, y) => dispatch(fetchParcelRequest(x, y)),
   onTransfer: parcel => dispatch(openModal('TransferModal', parcel))
 })
